% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_multiband_image.R
\name{create_multiband_image}
\alias{create_multiband_image}
\title{Create multiband image}
\usage{
create_multiband_image(mosaics_dir, desired_band_order)
}
\arguments{
\item{mosaics_dir}{Description of the parameter x.}

\item{desired_band_order}{description}
}
\value{
combined tif image
}
\description{
A description of what your function does.
}
