% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_optimum_thresholds.R
\name{find_optimum_thresholds}
\alias{find_optimum_thresholds}
\title{Find Optimum Thresholds}
\usage{
find_optimum_thresholds(df, class, band, site, class_value)
}
\arguments{
\item{df}{description}

\item{class}{name of column where class/catgegories are stored for}

\item{band}{description}

\item{site}{description}

\item{class_value}{description}
}
\value{
a df with optimum thresholds for each site
}
\description{
Takes input ground truth data to find optimum NDVI and NIR Thresholds for image masking. Input must be a df with column for band values and column indicating class (i.e. )
}
\examples{
ndvi_thresholds <- find_optimum_thresholds(ndvi_values, class = 'class', value = 'ndvi', site = 'site', class_value = 'veg')
}
