% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_spectral_metrics.R
\name{calculate_cv}
\alias{calculate_cv}
\title{Calculate spectral metrics}
\usage{
calculate_cv(
  pixel_values_df,
  wavelengths,
  rarefaction = FALSE,
  min_points = NULL,
  n = NULL
)
}
\arguments{
\item{pixel_values_df}{the data frame containing pixel values, obtained from extract_pixel_values function}

\item{wavelengths}{a list of wavelengths which correspond to column names from pixel_values_df}

\item{rarefaction}{either TRUE or FALSE to apply rarefaction step. substantially increases processing time}

\item{min_points}{if rarefaction = T, the minimum number of pixels for any aoi - to standardise uneven number of pixels across different sites}

\item{n}{if rarefaction = T, number of subset permutations}
}
\value{
a dataframe containing spectral metrics for each aois within each site/raster
}
\description{
Calculates CV, SV, and CHV from pixel values dataframe with columns for each wavelength + site_name + aoi_id
}
