% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_spectral_metrics.R
\name{calculate_cv}
\alias{calculate_cv}
\title{Calculate spectral metrics}
\usage{
calculate_cv(
  pixel_values_df,
  aoi_id,
  wavelengths,
  rarefaction = F,
  n = NULL,
  min_points = NULL
)
}
\arguments{
\item{df}{description}
}
\description{
Calculates continuous spectral metrics (CV, SV, CHV) based on input pixel value df
}
