% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_field_diversity.R
\name{calculate_field_diversity}
\alias{calculate_field_diversity}
\title{Calculate field diversity}
\usage{
calculate_field_diversity(survey_data)
}
\arguments{
\item{survey_data}{a df obtained from ausplotsR package with survey data for appropriate sites}
}
\value{
a list, containing 'field diversity' - values of species richness, shannon's index, simpsons index, exponential shannon's, inverse simpson, pielou's evenness, and 'community matrices'
}
\description{
Takes ausplot survey data, calculates field diversity for each site
}
\examples{
ausplot.pi.data <- ausplotsR::get_ausplots(my.Plot_IDs=
    c("SATFLB0004", "QDAMGD0022", "NTASTU0002"), veg.PI=TRUE)$veg.PI
field_diversity <- calculate_field_diversity(ausplot.pi.data)

}
