% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_masked_raster.R
\name{create_masked_raster}
\alias{create_masked_raster}
\title{Create Masked Raster}
\usage{
create_masked_raster(
  input,
  output_dir,
  ndvi_threshold = NULL,
  nir_threshold = NULL,
  ndvi_threshold_df = NULL,
  nir_threshold_df = NULL,
  red_band_index = 3,
  nir_band_index = 5
)
}
\arguments{
\item{input}{can be a directory containing multiple tif/envi files, string of files, or single file.}

\item{output_dir}{an output directory for the masked raster/s to be saved}

\item{ndvi_threshold}{NDVI threshold (values beneath this contain non-veg pixels and should be masked)}

\item{nir_threshold}{NIR threshold (values beneath this contain shadows and should be masked)}

\item{ndvi_threshold_df}{optional - a two columned df with 'site' col (site values must match first string of input file name) and  threshold values for each file}

\item{nir_threshold_df}{optional - a two columned df with 'site' col (site values must match first string of input file name) and  threshold values for each file}

\item{red_band_index}{layer number for red band}

\item{nir_band_index}{layer number for nir band}
}
\value{
A masked raster image, saved in the output directory
}
\description{
Creates a masked raster, removes non-vegetation pixels using input ndvi & nir threshold values
}
\examples{
input <- system.file("extdata/multiband_image", package = "saltbush")
output_dir <- tempdir()
create_masked_raster(input,
                    output_dir,
                    ndvi_threshold = 0.02,
                   nir_threshold = 0.04,
                     red_band_index = 3,
                     nir_band_index = 5)
}
