% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_multiband_image.R
\name{create_multiband_image}
\alias{create_multiband_image}
\title{Create multiband image}
\usage{
create_multiband_image(input_dir, desired_band_order, output_dir)
}
\arguments{
\item{desired_band_order}{Order of files to be combined. should be in wavelength order. provide list of file basenames in correct order}

\item{output_dir}{Folder to store combined tif file}

\item{mosaics_dir}{Folder which contains single band tifs to be combined}
}
\value{
combined tif file
}
\description{
Combines single reflectance band tifs into a multiband image.
}
\examples{
create_multiband_image('001_site_mosaics', c('blue', 'green', 'red', 'red_edge', 'nir'), 'data_out/001_site')
}
