% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_multiband_image.R
\name{create_multiband_image}
\alias{create_multiband_image}
\title{Create multiband image}
\usage{
create_multiband_image(input_dir, desired_band_order, output_dir)
}
\arguments{
\item{input_dir}{Directory containing folders with waveband tifs to be combined, or single folder with waveband tifs to be combined}

\item{desired_band_order}{Order of files to be combined. should be in wavelength order. provide vector of file basenames in correct order}

\item{output_dir}{Folder to store combined tif file}
}
\value{
combined tif file
}
\description{
Combines single reflectance band tifs into a multiband image.
}
\examples{
create_multiband_image('2024_band_rasters', c('blue', 'green', 'red', 'red_edge', 'nir'), 'data_out/2024_rasters')
}
