% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pixel_values.R
\name{extract_pixel_values}
\alias{extract_pixel_values}
\title{Extract pixel values}
\usage{
extract_pixel_values(raster_files, aoi_files, wavelength_names)
}
\arguments{
\item{raster_files}{directory of input raster files}

\item{aoi_files}{area of interest file - shapefile containing one or more site polygons for each raster}

\item{wavelength_names}{the wavelength corresponding to each layer of the raster_files}
}
\value{
a df with pixel values for each of the image layers
}
\description{
Extracts pixel values from each layer of multiband image and creates pixel value df
}
\examples{
aoi_files <- list.files('inst/extdata/aoi',
   pattern = 'image_aoi.shp$', full.names = TRUE)
raster_files <- list.files('inst/extdata/multiband_image',
   pattern = '.tif$', full.names = TRUE)
pixelvalues <- extract_pixel_values(raster_files, aoi_files, c('blue','green','red','red_edge','nir'))
}
