% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pixel_values.R
\name{extract_pixel_values}
\alias{extract_pixel_values}
\title{Extract pixel values}
\usage{
extract_pixel_values(raster_files, aoi_files, wavelength_names)
}
\arguments{
\item{raster_files}{directory of input raster files}

\item{aoi_files}{area of interest file - shapefile containing one or more site polygons for each raster}

\item{wavelength_names}{wavelength names for each band, must match order of stacked layers}
}
\value{
a df with pixel values for each of the image layers
}
\description{
Extracts pixel values from each layer of multiband image and creates pixel value df
}
\examples{
set.seed(123)
df <- data.frame(
site_name = rep(c("site_one", "site_two", "site_three", "site_four"), each = 5000),
aoi_id = 1,
blue = runif(20000, min = 0, max = 1),
green = runif(20000, min = 0, max = 1),
red = runif(20000, min = 0, max = 1),
red_edge = runif(20000, min = 0, max = 1),
nir = runif(20000, min = 0, max = 1))
pixelvalues <- calculate_cv(df,
wavelengths = c('blue','green','red','red_edge','nir'),
rarefaction = TRUE, min_points = 5000, n = 999)
}
