% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_optimum_thresholds.R
\name{find_optimum_thresholds}
\alias{find_optimum_thresholds}
\title{Find Optimum Thresholds}
\usage{
find_optimum_thresholds(
  df,
  class_col,
  band_or_index_col,
  site_col,
  class_value
)
}
\arguments{
\item{df}{Input must be a df with columns for site name, pixel value, class}

\item{class_col}{name of column where object class are stored}

\item{band_or_index_col}{name of column where pixel values are stored - e.g. 'ndvi' or 'nir_values'}

\item{site_col}{name of site column - which indicates where pixel values + class data are obtained from}

\item{class_value}{value/s of class that should be masked - e.g. for ndvi, 'non-veg' or c('bare_ground','rocks), for nir, 'shadow'}
}
\value{
a df with optimum thresholds for each site
}
\description{
Takes input ground truth data to find optimum thresholds for image masking.
}
\examples{
ndvi_thresholds <- find_optimum_thresholds(ndvi_values, class = 'class', value = 'nir', site = 'site', class_value = 'shadow')
}
